Refactor: Extract the interrupt handling into DataProvider.
